+++ OpenBullet2.Web/Dtos/Wordlist/WordlistDto.cs
@@ -30,7 +30,7 @@ public class WordlistDto
     /// <summary>
     /// The total number of lines in the wordlist.
     /// </summary>
-    public int LineCount { get; set; }
+    public long LineCount { get; set; }
 
     /// <summary>
     /// The wordlist type.
